rm moDllz
For unknown reasons, this repo confuses submodules
into thinking it is always changed and shows
up as 'modified content'.

(I suspect it may be related to line endings
settings and somehow the svgs get modified
on pull, but haven't tracked it down)

Instead of working around it when updating, just
removing.